Only call log() on float or double types.
This works around a problem with Solaris' libc where log() in C++ is
overloaded for float, double and long double but not integer types.

Jeroen Demeyer, 25 May 2012

+++ QS.cpp
@@ -467,7 +467,7 @@ void computeSizes(unsigned long numPrimes)
      primeSizes = (unsigned char *) calloc(sizeof(unsigned char),numPrimes);
      for (unsigned long i = 0; i<numPrimes; i++)
      {
-         primeSizes[i]=(unsigned char)floor(log(factorBase[i])/log(2.0)-FUDGE+0.5);
+         primeSizes[i]=(unsigned char)floor(log((double)factorBase[i])/log(2.0)-FUDGE+0.5);
      }
      
      return;
